NCL-7411 Rename PageTitles file to constants
